+++ CHANGES.rst
@@ -4,6 +4,8 @@
 - Dropped support for Python 3.6 and 3.7. Minimum supported Python
   version is now 3.8. [#330]
 
+- OBMAG and VEGAMAG are no longer interchangeable. [#331]
+
 1.1.1 (2021-11-18)
 ==================